requirements.txt
alembic==1.8.0
cffi==1.15.1
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
flask-swagger-ui==4.11.1
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
numpy==1.23.1
pandas==1.4.3
pycparser==2.21
PyJWT==2.4.0
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.39
Werkzeug==2.1.2
WTForms==3.0.1
